Removed escript file as option
It is gernally better to just use the parse transform. Using the
escript still required rec2json as a dependancy, and using raw
records to make modules just meant they were harder to extend later.
Further, the script increased the steps in a compile. Using a parse
transform is just plain better.