Fixed bug for a field typed with a list of records
Given the definition:

-record(inner, {f}).
-record(outer, {f :: [#inner{}]}).

The erlang compiler created a type spec for outer.f as:
{type, L1, list, [{type, L2, record, [{atom, L3, RecordName}]}]}

Most other types, the record type tuple does not have the record name
in a list. Thus, then the r2j_compiler came upon it, in maintained the list.
The rec2json, when doing a from_json, would hit upon the record type, and
assume the second element was a name, and not a list.

r2j_compile now detects this edge case, and turns it into the expected
{record, Name} for rec2json.